Extract dashboard widget renderers
