+++ dev-docs/architecture.md
@@ -57,7 +57,8 @@ js/
   lens-pages.js     — Labs, Genome, Body, Insight, and Recommendations page renderers
   dashboard-widgets.js — dashboard widget registry, defaults, widget prefs
   dashboard-widget-controls.js — widget picker, layout actions, drag/reorder controls
-  views.js          — dashboard renderers, Light/category views, modals, manual entry, onboarding
+  dashboard-widget-renderers.js — dashboard widget body renderers and recommendation helpers
+  views.js          — dashboard composition, Light/category views, modals, manual entry, onboarding
   crypto.js         — AES-256-GCM encryption, cross-tab sync (BroadcastChannel)
   backup.js          — IndexedDB auto-backup, folder backup (File System Access API), backup restore
   lab-context.js     — buildLabContext() central AI context serializer (extracted from chat.js)
@@ -125,8 +126,9 @@ Modules in a higher layer may import from lower layers. Modules in the same laye
                                  ▼
 ┌─────────────────────────────────────────────────────────────────────┐
 │  L6 — Orchestration                                                 │
-│  dashboard-widgets.js  dashboard-widget-controls.js  views.js      │
-│  main.js  tour.js  changelog.js                                    │
+│  dashboard-widgets.js  dashboard-widget-controls.js                 │
+│  dashboard-widget-renderers.js  views.js  main.js  tour.js         │
+│  changelog.js                                                       │
 └─────────────────────────────────────────────────────────────────────┘
 ```
