+++ tests/test-correctness-phase2.js
@@ -79,6 +79,7 @@ const pwaAppShellAssets = [
   '/js/lens-pages.js',
   '/js/dashboard-widgets.js',
   '/js/dashboard-widget-controls.js',
+  '/js/dashboard-widget-renderers.js',
   '/js/light-devices.js',
   '/js/light-env.js',
   '/js/light-tools.js',
